couldn't resist adding a sherlock holmes quote.
also, decided to point the docs link just to the root page, rather
than into the reference manual. (reference.md manual is more complete
w.r.t. `match` than e.g. trpl is, but long-term that will hopefully
not be the case, and so over time I expect the reference.md to be the
wrong thing to have long-lived links to.)