rust-encoding is not "Servo’s"
The phrasing suggested that Servo "owns" the rust-encoding library. This is not the case: it is primarily developed by Kang Seonghoon who is a member of the Rust community but not directly affiliated to Servo.

To avoid sounding like Servo is stealing credit from Kang, this changes the example to rust-url which I developed while employed by Mozilla to work on Servo, and is also used in a number of projects outside of Servo.