Improve site legibility
This patch removes unnecessary letter-spacing and line-height settings,
as well as the forced Arial/Helvetica font override, an instead uses the
user's chosen font and the font's settings for letter-spacing and
line-height. It also removes the global override for initial page font
size, so that the user's font size settings are preserved.

The blockquote styles are also update so that the quoted text is
formatted similarly to the surrounding text (no italics, and no larger
font size), and with a blacker grey to make the text more legible.