Fix profile directory in crate graph self-profiling example
In the rustc self-profiling blog post, the proposed crate graph-wide self-profiling command is incorrect.

```sh
$ RUSTFLAGS="-Zself-profile=./profiles -Zself-profile-events=default,args" cargo build
```

The problem is that the `./profiles` relative directory will be resolved with respect to whatever directory rustc happens to be working in at the time where the RUSTFLAGS are parsed. Which, for dependency crates, will be the source directory of those dependencies, not the source directory of the crate that we are working on.

Forcing directory resolution to be relative to the active crate's directory by replacing the rustc-evaluated `.` with a shell-evaluated `$(pwd)` resolves this problem.

Kudos to @bjorn3 for figuring this out!