Do blog generation as part of the invocation of the embedded web server.
(I'm not yet trying to do any fancy monitoring of changes to the files and
subsequent regeneration. This is all meant to be very bare bones.)

To accomplish this I had to make the main crate callable from other crates,
which in turn required some small changes to how it pulls in its own child
modules, since `main` can no longer assume it is the crate root.