Fixed dead link in posts/2023-08-29-committing-lockfiles.md
Said blog post contains a link to the Cargo book. The heading that was being
linked to was changed from "Multiple requirements" to "Multiple version
requirements" (in [this commit][doc-change]), this change was not reflected in
the blog post. I am not too sure whether the link has ever worked as the docs
were changed on 2023-08-27 while the blog post was only written on 2023-08-29,
but either way, this commit fixes the link.

[doc-change]: https://github.com/rust-lang/cargo/commit/01e1ff0484b62db28f0a7e231a3a7ad5f7439bca?diff=unified&w=0#diff-d184a3240b09684abaca1a50dc8519dab2a9f5a17631180036f4b8090bb8a2d0L105