Remove automatic `$` prefix for `bash` and `console` code blocks
This prefix is very surprising to blog authors because e.g. GitHub does not render it that way. It also does not work well if a code block contains multiple commands where all but the first one need a manual prefix. The original commit that introduced this unfortunately does not give any reasons for why it was introduced.