Revise the Rust 2024 call for testing
Let's do a bit more to get people excited, and tighten up some
language here.

For the team in the by-line, it'd be best if we had a website for our
team, but for the moment, let's just link to the edition guide rather
than to the team TOML file.