Add system theme setting
The system setting is followed if there is no item
`blog-rust-lang-org-theme` in local storage. In order to switch back to
the system theme, that item is deleted.

Previously, if a system preference for dark theme was detected, that
was additionally stored explicitly as `blog-rust-lang-org-theme`.
That was removed, because it prevents the system setting from being
followed if the user switches their system back to light theme.