Avoid theme flicker if system setting is dark
Previously, `@media (prefers-color-scheme: dark)` was only used in the
`noscript` stylesheet, but that's the best way to immediately set the
dark theme without any flickering. That is now integrated into the the
main stylesheet.

The flickering is still present if the user has a default system setting
of light mode but explicitly set it to dark. Actually, this introduces
flickering from dark to light if the user has a system setting of dark
but an explicit override of light. So it goes both ways. I think this
is a good tradeoff, most users will want the website to just follow the
system setting.

The `theme-switch.js` can also be simplified a little, since it's not
necessary anymore to set the `data-theme` attribute if there is no
explicit override.