DO NOT MERGE - Add snapshot tests for large refactors
This patch adds snapshot testing for essentially every file that's
generated as part of the website. This should not be merged, because
The number of snapshot files is massive and a new one would have to be
added for every new post.

However, this can be very useful to test refactors of the code that
shouldn't change the output in any way. (Or if it does, these tests
make it easier to see if the changes to the output are appropriate.)

The process of using these tests would go something as follows:
- install [cargo-insta](https://insta.rs/docs/quickstart/)
- create a new branch, call it for example my-foo-refactor-testing-base
- cherry pick this commit onto that branch
- run `cargo insta test --accept` to generate the baseline snapshots
- commit all these files and push the branch (it will not be merged)
- create a new branch, e.g. my-foo-refactor, based on *-testing-base
- do your refactoring
- commit and push your changes, and create a PR against the *-testing-base branch

Now you have a nice PR that's fully snapshot tested and can go through
regular review cycles. Before merging, interactively rebase your
refactor branch, drop the snapshot-related commits and re-target the PR
to master.