Add front_matter crate
We might want to move the blog to Zola. In that case, it will still be
useful to have a crate that validates and even fixes our front matter.

This will be especially important if we decide to duplicate the date of
a post in the two front matter fields `date` and `path`, which may be
necessary to preserve the existing URL scheme.

While the custom static site generator is still present, it can reuse
the front matter parsing logic.