Mention AdaCore regarding FLS and strike "donation"
There was some relationship between Ferrous Systems and AdaCore when
they worked together to create the FLS.  Apparently that relationship
ceased in July 2023, according to the Foundation's blog post, but I
don't think we need to go into those details.  Still, it seems worth
mentioning AdaCore here, regarding the creation, so let's do that.

In discussion with various people, there's some preference to not use
the word "donation".  It's awkward, then, that word is part of the
headline on the Foundation's post which has now been widely quoted and
reported on, so it seems too late to change that.  But the Foundation
has struck that word from the remainder of their post.

In our case, we weren't leaning on the word heavily anyway.  Looking
at it from a Project perspective, our primary focus is that we're
"adopting" the FLS, and that's still right of course.  We'd used the
word "donation" in a more limited way to contextualize what we're
thanking Ferrous Systems for doing.  But looking at the text now, we
can without any real loss surgically strike that word and replace it
with... nothing.  So let's do that.