Make front matter test more lenient
During the migration to Zola, it was nice to have the enforcement that
all front matter is formatted identically. But going into the future,
the purpose of this test will be to help blog authors spot mistakes in
their front matter. It would be annoying noise a fail because the keys
are in a different order, whitespace, comments etc.