Reduce risk of publishing posts with wrong date
The date of publication often has to be adjusted at the last minute.
This can easily be forgotten. In order to reduce the risk of posts with
an incorrect date being published, this commit encourages authors to
keep a placeholder date until shortly before publication. A CI check
prevents the deployment if a placeholder date is detected.

The placeholder date "9999/12/31" is chosen because it shows up at the
top of the list of posts during development and it's clearly "invalid".