Link to edit page directly (#1607)
Contributors may click on the "fix a typo" link while reading a specific
post. When that only leads to the repo, it takes extra effort to
find the specific file to edit. Linking to that file directly is more
convenient. Even if contributors don't want to use the GitHub editor, it
at least tells them the exact path of the file in the repo.