Fix year delimiters in releases page
The previous approach to detect if a release is the last one in a year
was incorrect. It checked if the following post was in the next year.
However, the following post doesn't have to be a release, so the check
could fail.

The result was missing year delimiters "Releases in YYYY".

A mutable variable makes it easier to track the actual year of the last
rendered release post.