Revise editorially
Let's clean up whitespace and capitalize the "Rust Project".

Let's move toward putting the year in the title -- we mean to do this
for many years, so it'll make more sense to be unambiguous.  Let's
also put the date last, as it's just metadata in some sense.

Most of our blog posts are titled in sentence case.  Let's move to
doing that here as well.