Recommend to compile zola in debug mode
Zola takes a long time to compile. For blog authors, it's probably
better to compile zola a little faster while sacrificing some
performance when building the site.

By default, zola rebuilds every page when something changes. Debug
builds perform noticeably worse in that case. But with the `--fast`
flag, zola only rebuilds pages that actually changed. With that, the
performance of debug and release builds not not noticeably different.