Fix claim about Git's default branch name
The blog post about changing the primary branch name of
`rust-lang/rust`, as written, claims that the default branch name
of *Git*, the open-source version control tool originally written by
Linus Torvalds, has changed.  As best I can determine, that's not
true, at least as of Git v2.51.  The mailing list thread from 2020,
linked in the post, certainly does not support this claim.  That
thread discusses making the default branch name for newly-created
repositories configurable, and it causes a warning to be shown if this
is not configured (indicating that the default may change in the
future), but the default when not configured has not been changed.

Conversely, *GitHub*, the service now owned by Microsoft, did change
the default branch name for newly-created repositories on 2020-10-01.
Probably this is what we mean instead.  Let's change the post to point
to that.